# backend/api_specification.yaml
openapi: 3.0.3
info:
  title: Chikitsa HMS API
  description: Hospital Management System API for managing patients, doctors, appointments, and medical records
  version: 1.0.0
  contact:
    name: Chikitsa Support
    email: support@chikitsa.com
servers:
  - url: http://localhost:5000
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [admin, doctor, patient]
    LoginResponse:
      type: object
      properties:
        status:
          type: string
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            user:
              type: object
              properties:
                id:
                  type: integer
                username:
                  type: string
                role:
                  type: string
    RegisterPatientRequest:
      type: object
      required: [username, email, password, first_name, last_name, dob, gender, phone]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        first_name:
          type: string
        last_name:
          type: string
        dob:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        blood_group:
          type: string
        phone:
          type: string
          pattern: '^\d{10}$'
        address:
          type: string
        emergency_contact_name:
          type: string
        emergency_contact_phone:
          type: string
        medical_history:
          type: string
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
  /auth/register/patient:
    post:
      tags: [Auth]
      summary: Register new patient
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPatientRequest'
      responses:
        '201':
          description: Registration successful
        '400':
          description: Validation error
...
  /auth/profile:
    get:
      tags: [Auth]
      summary: Get user profile
      responses:
        '200':
          description: Profile retrieved
    put:
      tags: [Auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Profile updated

  # ==================== ADMIN - DASHBOARD ====================
...